backend/python/requirements.txt
scapy>=2.5.0
kafka-python>=2.0.2
flask>=3.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
